Remove unused random value condition and fix ACK frame not being accounted.
Procedure queue_frame had a test to use a fixed random value that was removed,
because it was not being used.
This same procedure had a bug that impeded the ACK frame transmission time
from being accounted in the transmission time of a frame, when the frame
transmission was successful. Previously the ACK transmission time was only
being accounted when the transmission failed.