Fix simultaneous transmissions in a medium.
Frames with different priorities could end up being transmitted
at the same time, if a frame of lower priority started being
transmitted and a new frame with higher priority arrived. This
happened because the search for frames was limited to the frames
with same or higher priority. The error was at the end of queue_frame.